Updated tracing pool implementation to be in line with comments
