+++ projects/rocprofiler-sdk/source/lib/rocprofiler-sdk/tracing/tracing_data_pool.hpp
@@ -25,177 +25,17 @@
 #include "fwd.hpp"
 
 #include <memory>
-#include <vector>
 
 namespace rocprofiler
 {
 namespace tracing
 {
-namespace  // Anonymous namespace
-{
-// Internal pool implementation
-class pool_impl
-{
-public:
-    pool_impl()
-    {
-        // Pre-allocate some objects to reduce initial allocation overhead
-        pool_.reserve(8);
-    }
+using pooled_tracing_data_t = std::unique_ptr<tracing_data, void (*)(tracing_data*)>;
 
-    ~pool_impl() = default;
-
-    // Get a tracing_data object (reused or new)
-    tracing_data* acquire()
-    {
-        if(!pool_.empty())
-        {
-            auto* obj = pool_.back().release();
-            pool_.pop_back();
-            return obj;
-        }
-        return new tracing_data{};
-    }
-
-    // Return object to pool
-    void release(tracing_data* obj)
-    {
-        if(!obj) return;
-
-        // Clear the object for reuse (keeps capacity allocated)
-        obj->callback_contexts.clear();
-        obj->buffered_contexts.clear();
-        obj->external_correlation_ids.clear();
-
-        // Prevent pool from growing unbounded
-        if(pool_.size() < 16)
-        {
-            pool_.emplace_back(obj);
-        }
-        else
-        {
-            delete obj;
-        }
-    }
-
-private:
-    std::vector<std::unique_ptr<tracing_data>> pool_;
-};
-
-// Thread-local pool instance
-// Returns nullptr if pool has been destroyed (during thread shutdown)
-inline pool_impl*
-get_pool()
-{
-    // Track if the pool is still valid
-    thread_local bool      pool_destroyed = false;
-    thread_local pool_impl pool;
-
-    // Use a destructor guard to detect when thread locals are being destroyed
-    struct pool_guard
-    {
-        bool& destroyed;
-        pool_guard(bool& d)
-        : destroyed(d)
-        {}
-        ~pool_guard() { destroyed = true; }
-    };
-    thread_local pool_guard guard(pool_destroyed);
-
-    // Return nullptr if pool is being/has been destroyed
-    if(pool_destroyed) return nullptr;
-
-    return &pool;
-}
-
-}  // anonymous namespace
-
-// RAII wrapper for automatic return to pool - PUBLIC interface
-class pooled_tracing_data
-{
-public:
-    pooled_tracing_data()
-    {
-        auto* pool = get_pool();
-        if(pool)
-        {
-            data_   = pool->acquire();
-            pooled_ = true;
-        }
-        else
-        {
-            data_   = new tracing_data{};
-            pooled_ = false;
-        }
-    }
-
-    ~pooled_tracing_data()
-    {
-        if(pooled_)
-        {
-            auto* pool = get_pool();
-            if(pool)
-                pool->release(data_);
-            else
-                delete data_;  // Pool destroyed, just delete
-        }
-        else
-        {
-            delete data_;
-        }
-    }
-
-    // No copy, but allow move
-    pooled_tracing_data(const pooled_tracing_data&) = delete;
-    pooled_tracing_data& operator=(const pooled_tracing_data&) = delete;
-
-    pooled_tracing_data(pooled_tracing_data&& other) noexcept
-    : data_(other.data_)
-    , pooled_(other.pooled_)
-    {
-        other.data_ = nullptr;
-    }
-
-    pooled_tracing_data& operator=(pooled_tracing_data&& other) noexcept
-    {
-        if(this == &other) return *this;
-
-        // Release current resource back to the pool
-        if(data_ != nullptr)
-        {
-            if(pooled_)
-            {
-                auto* pool = get_pool();
-                if(pool)
-                    pool->release(data_);
-                else
-                    delete data_;
-            }
-            else
-            {
-                delete data_;
-            }
-        }
-
-        // Steal the resource from other
-        data_       = other.data_;
-        pooled_     = other.pooled_;
-        other.data_ = nullptr;
-
-        return *this;
-    }
-    // Access the tracing_data
-    tracing_data&       operator*() { return *data_; }
-    tracing_data*       operator->() { return data_; }
-    tracing_data*       get() { return data_; }
-    const tracing_data& operator*() const { return *data_; }
-    const tracing_data* operator->() const { return data_; }
-    const tracing_data* get() const { return data_; }
-
-private:
-    tracing_data* data_;
-    bool          pooled_;  // Track whether this instance uses the pool
-};
+// Allocate a tracing_data object from thread-local pool
+// Returns a unique_ptr that automatically returns the object to pool on destruction
+pooled_tracing_data_t
+allocate_tracing_data();
 
 }  // namespace tracing
 }  // namespace rocprofiler
